python/requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
Flask-CORS
python-dotenv